// libs/parser/src/ast/node-type.ts
export enum AstNodeType {
  ExprStmt = 'ExprStmt',
  BasicLit = 'BasicLit',
  Identifier = 'Identifier',
  SelectorExpr = 'SelectorExpr',
  IndexExpr = 'IndexExpr',
  CallExpr = 'CallExpr',
  ScopeResolver = 'ScopeResolver',
  UnaryExpr = 'UnaryExpr',
  BinaryExpr = 'BinaryExpr',
  TernaryExpr = 'TernaryExpr',
  ParenExpr = 'ParenExpr',
  JsonRoot = 'json',
  JsonObject = 'jsonObject',
  JsonArray = 'jsonArray',
  JsonProperty = 'jsonProperty',
  JsonLiteral = 'jsonLiteral',
  StringTemplate = 'stringTemplate',
}